Revert "fix(MToken): fix mint silent overflow"
This reverts commit 312890ce5bf43cc784553a3841f0d1a2028724b9.

+++ src/MToken.sol
@@ -225,10 +225,8 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
             _addNonEarningAmount(recipient_, UIntMath.safe240(amount_));
         }
 
-        // NOTE: Need to cast to uint256 to avoid overflowing uint112 before the comparison.
         if (
-            uint256(principalOfTotalEarningSupply) + _getPrincipalAmountRoundedDown(totalNonEarningSupply) >=
-            type(uint112).max
+            principalOfTotalEarningSupply + _getPrincipalAmountRoundedDown(totalNonEarningSupply) >= type(uint112).max
         ) {
             revert OverflowsPrincipalOfTotalSupply();
         }
